Added static_cast

+++ Straight_skeleton_2/test/Straight_skeleton_2/test_sls.cpp
@@ -766,7 +766,7 @@ int test_zone ( Zone& rZone )
       {
         if ( sOffsetAtEntry.size() > 0 )
         {
-          int lSize = std::distance(lTimes.begin(),lTimes.end());
+          int lSize = static_cast<int>(std::distance(lTimes.begin(),lTimes.end()));
           
           for ( std::vector<int>::const_iterator oi = sOffsetAtEntry.begin() ; oi != sOffsetAtEntry.end() ; ++ oi )
           {
